move LookupItemVmJsonConverter to new place
